February 2025 Pre-Event Second Pass
Filling out the information for the game section of the Feb 2025 event page.
Editing the CSS for the dev-social class to add the display: contents rule, allowing multiple social buttons/icons to be displayed horizontally in the dev section of a project card.
Also adding a rule to the .dev-social a element that removes text-decoration. this removes a blue underline near the buttons